fix(playerbot): Move ObjectiveAnalytics to interface
- Moved ObjectiveAnalytics struct to IObjectiveTracker.h with full definition
- Added atomic and chrono includes to interface
- Removed duplicate from ObjectiveTracker.h
- All objective tracking structs now in interface headers
